Added NER feature
