+++ data/corpus-heroes.json
@@ -1,0 +1,28 @@
+{
+	"name": "Corpus-Heroes",
+	"locale": "en-US",
+	"contextData": "./data/heros.json",
+	"data": [
+		{
+			"intent": "hero.realname",
+			"utterances": ["what is the real name of @hero", "who is @hero"],
+			"answers": [
+				"The real name of {{ hero }} is {{ _data[entities.hero.option].realName }}"
+			]
+		},
+		{
+			"intent": "hero.city",
+			"utterances": ["where @hero lives?", "what's the city of @hero?"],
+			"answers": ["{{ hero }} lives at {{ _data[entities.hero.option].city }}"]
+		}
+	],
+	"entities": {
+		"hero": {
+			"options": {
+				"spiderman": ["spiderman", "spider-man"],
+				"ironman": ["ironman", "iron-man"],
+				"thor": ["thor"]
+			}
+		}
+	}
+}
