+++ data/heros.json
@@ -1,0 +1,14 @@
+{
+	"spiderman": {
+		"realName": "Peter Parker",
+		"city": "Queens, New York"
+	},
+	"ironman": {
+		"realName": "Tony Stark",
+		"city": "Stark Tower, New York"
+	},
+	"thor": {
+		"realName": "Odinson",
+		"city": "Asgard"
+	}
+}
