--- notes/intent-template.json
@@ -1,4 +1,0 @@
-{
-	"questions": [],
-	"answers": []
-}
