+++ package.json
@@ -10,6 +10,7 @@
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
+		"@nlpjs/basic": "^4.22.9",
 		"ejs": "^3.1.6",
 		"express": "^4.17.2",
 		"node-nlp": "^4.22.9",
